+++ DaftMobileTask/Assets/_Project/Scripts/Persistance/GameStatePersistance.cs
@@ -5,14 +5,14 @@ public static class GameStatePersistance
 {
     public static void SaveGameStateData(GameState gameState)
     {
-        PlayerPrefs.SetFloat("BestTimeScore", gameState.BestTimeScore);
+        PlayerPrefs.SetInt("BestTimeScore", gameState.BestTimeScore);
     }
 
     public static GameState LoadGameStateData()
     {
         GameState gameState = new GameState();
 
-        gameState.BestTimeScore = PlayerPrefs.GetFloat("BestTimeScore", 0);
+        gameState.BestTimeScore = PlayerPrefs.GetInt("BestTimeScore", 0);
 
         return gameState;
     }
